+++ neoforge/src/main/java/dev/tocraft/ctgen/neoforge/CTGNeoForgeEventListener.java
@@ -8,7 +8,7 @@ import dev.tocraft.ctgen.worldgen.MapBasedBiomeSource;
 import dev.tocraft.ctgen.worldgen.MapBasedChunkGenerator;
 import dev.tocraft.ctgen.worldgen.noise.CTGAboveSurfaceCondition;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.server.packs.resources.PreparableReloadListener;
+import net.minecraft.world.level.biome.Biome;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.neoforge.common.NeoForge;
 import net.neoforged.neoforge.event.AddReloadListenerEvent;
@@ -17,9 +17,6 @@ import net.neoforged.neoforge.network.event.RegisterPayloadHandlersEvent;
 import net.neoforged.neoforge.registries.RegisterEvent;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-import java.util.concurrent.CopyOnWriteArrayList;
 
 @ApiStatus.Internal
 public final class CTGNeoForgeEventListener {
@@ -32,17 +29,9 @@ public final class CTGNeoForgeEventListener {
         modEventBus.addListener(CTGNeoForgeEventListener::registerPayload);
     }
 
-    private static final List<PreparableReloadListener> RELOAD_LISTENERS = new CopyOnWriteArrayList<>() {
-        {
-            add(new BiomeImageRegistry());
-            add(new HeightImageRegistry());
-        }
-    };
-
-    private static void addReloadListenerEvent(AddReloadListenerEvent event) {
-        for (PreparableReloadListener reloadListener : RELOAD_LISTENERS) {
-            event.addListener(reloadListener);
-        }
+    private static void addReloadListenerEvent(@NotNull AddReloadListenerEvent event) {
+        event.addListener(new BiomeImageRegistry());
+        event.addListener(new HeightImageRegistry());
     }
 
     private static void registerCommands(@NotNull RegisterCommandsEvent event) {